Update submission.js
declared variables for a boolean value, a string and an array.

for loop to iterate through passed param 'str' from end to beginning and set a new string as the reverse order of the string indexes. 

test the original param 'str' against the reversed index order string.  Boolean value set to true on a match, else false.  

returned boolean value