Update submission.js
number variables 'max' and 'min' are declared and instantiated with values of Math.max and Math.min from the 'counted' array of numbers.

for loop used to iterate over object array 'countArray' to compare 'max' and 'min' 'counted' variables to each 'countArray' object count property value.  
The correlating letter value of the object with the 'count' value equal to the 'max' and 'min' 'counted' array values are assigned to the 'mostLeast' object property values of 'most' and 'least' assigned to letter value of object with count value equal to objects with respective count values of max and min.